gcc 4 fixes
git-svn-id: https://svn.aros.org/svn/aros/trunk@23844 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/fish/sunmaze/maze.h
@@ -1,7 +1,7 @@
 #include <proto/alib.h>
 
-#define StackX (short) *(stack+(stackptr<<1))
-#define StackY (short) *(stack+(stackptr<<1)+1)
+#define StackX *(stack+(stackptr<<1))
+#define StackY *(stack+(stackptr<<1)+1)
 
 /* MAZETEST (xpos,ypos,mazewidth,pointer to buffer) */
 #ifdef BITMAP
